First cut at Hello Datawire Connect. JAVASCRIPT ONLY SO FAR. INCLUDES BRUTAL HACK.
Hello Datawire Connect is helloRPC, but using Datawire Connect for service discovery. JavaScript only right now, Python to follow this afternoon.

The brutal hack is a race condition: the client will try to issue the RPC before it has a chance to get the routing table from Datawire Connect. Right now, that's "fixed" with a setTimeout, which is obviously brutally hackish and wrong.